+++ src/methods/create_chat_invite_link.rs
@@ -9,7 +9,7 @@ pub struct CreateChatInviteLink<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Point in time (Unix timestamp) when the link will expire
     #[serde(skip_serializing_if = "Option::is_none")]
     pub expire_date: Option<i32>,
@@ -19,7 +19,7 @@ pub struct CreateChatInviteLink<'a> {
 }
 
 impl<'a> CreateChatInviteLink<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64) -> Self {
         Self {
             chat_id: chat_id,
             expire_date: None,
@@ -32,7 +32,7 @@ impl<'a> CreateChatInviteLink<'a> {
         let resp = self.bot.send("createChatInviteLink", Some(string)).await?;
         Ok(serde_json::from_str::<ChatInviteLink>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
