+++ src/methods/delete_message.rs
@@ -9,13 +9,13 @@ pub struct DeleteMessage<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Identifier of the message to delete
     pub message_id: i32,
 }
 
 impl<'a> DeleteMessage<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, message_id: i32) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, message_id: i32) -> Self {
         Self {
             chat_id: chat_id,
             message_id: message_id,
@@ -27,7 +27,7 @@ impl<'a> DeleteMessage<'a> {
         let resp = self.bot.send("deleteMessage", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
