+++ src/methods/get_chat_administrators.rs
@@ -9,11 +9,11 @@ pub struct GetChatAdministrators<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target supergroup or channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
 }
 
 impl<'a> GetChatAdministrators<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64) -> Self {
         Self {
             chat_id: chat_id,
             bot: bot,
@@ -26,7 +26,7 @@ impl<'a> GetChatAdministrators<'a> {
             &resp.text().await?,
         )?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
