+++ src/methods/send_voice.rs
@@ -11,7 +11,7 @@ pub struct SendVoice<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Audio file to send. Pass a file_id as String to send a file that exists on the Telegram servers (recommended), pass an HTTP URL as a String for Telegram to get a file from the Internet, or upload a new one using multipart/form-data. More info on Sending Files
     pub voice: InputFile,
     /// Voice message caption, 0-1024 characters after entities parsing
@@ -41,7 +41,7 @@ pub struct SendVoice<'a> {
 }
 
 impl<'a> SendVoice<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, voice: InputFile) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, voice: InputFile) -> Self {
         Self {
             chat_id: chat_id,
             voice: voice,
@@ -61,7 +61,7 @@ impl<'a> SendVoice<'a> {
         let resp = self.bot.send("sendVoice", Some(string)).await?;
         Ok(serde_json::from_str::<Message>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
