+++ src/methods/set_chat_permissions.rs
@@ -9,13 +9,13 @@ pub struct SetChatPermissions<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target supergroup (in the format @supergroupusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// New default chat permissions
     pub permissions: ChatPermissions,
 }
 
 impl<'a> SetChatPermissions<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, permissions: ChatPermissions) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, permissions: ChatPermissions) -> Self {
         Self {
             chat_id: chat_id,
             permissions: permissions,
@@ -27,7 +27,7 @@ impl<'a> SetChatPermissions<'a> {
         let resp = self.bot.send("setChatPermissions", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
