+++ src/methods/set_chat_sticker_set.rs
@@ -9,13 +9,13 @@ pub struct SetChatStickerSet<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target supergroup (in the format @supergroupusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Name of the sticker set to be set as the group sticker set
     pub sticker_set_name: String,
 }
 
 impl<'a> SetChatStickerSet<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, sticker_set_name: String) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, sticker_set_name: String) -> Self {
         Self {
             chat_id: chat_id,
             sticker_set_name: sticker_set_name,
@@ -27,7 +27,7 @@ impl<'a> SetChatStickerSet<'a> {
         let resp = self.bot.send("setChatStickerSet", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
