+++ src/methods/set_game_score.rs
@@ -20,7 +20,7 @@ pub struct SetGameScore<'a> {
     pub disable_edit_message: Option<bool>,
     /// Required if inline_message_id is not specified. Unique identifier for the target chat
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub chat_id: Option<i32>,
+    pub chat_id: Option<i64>,
     /// Required if inline_message_id is not specified. Identifier of the sent message
     #[serde(skip_serializing_if = "Option::is_none")]
     pub message_id: Option<i32>,
@@ -63,7 +63,7 @@ impl<'a> SetGameScore<'a> {
         self.disable_edit_message = Some(disable_edit_message);
         self
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = Some(chat_id);
         self
     }
